+++ packages/cli-tools/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hot-updater/cli-tools
 
+## 0.21.10
+
+### Patch Changes
+
+- @hot-updater/plugin-core@0.21.10
+
+## 0.21.9
+
+### Patch Changes
+
+- aa399a6: chore: deps picocolors
+- Updated dependencies [aa399a6]
+  - @hot-updater/plugin-core@0.21.9
+
 ## 0.21.8
 
 ### Patch Changes