+++ packages/hot-updater/CHANGELOG.md
@@ -1,5 +1,37 @@
 # hot-updater
 
+## 0.21.10
+
+### Patch Changes
+
+- Updated dependencies [5289b17]
+  - @hot-updater/server@0.21.10
+  - @hot-updater/cli-tools@0.21.10
+  - @hot-updater/aws@0.21.10
+  - @hot-updater/cloudflare@0.21.10
+  - @hot-updater/firebase@0.21.10
+  - @hot-updater/plugin-core@0.21.10
+  - @hot-updater/console@0.21.10
+  - @hot-updater/core@0.21.10
+  - @hot-updater/supabase@0.21.10
+
+## 0.21.9
+
+### Patch Changes
+
+- 396ae54: feat(cli): db generate --sql create only sql
+- aa399a6: chore: deps picocolors
+- Updated dependencies [aa399a6]
+  - @hot-updater/plugin-core@0.21.9
+  - @hot-updater/cli-tools@0.21.9
+  - @hot-updater/console@0.21.9
+  - @hot-updater/server@0.21.9
+  - @hot-updater/aws@0.21.9
+  - @hot-updater/cloudflare@0.21.9
+  - @hot-updater/firebase@0.21.9
+  - @hot-updater/supabase@0.21.9
+  - @hot-updater/core@0.21.9
+
 ## 0.21.8
 
 ### Patch Changes