+++ plugins/js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hot-updater/js
 
+## 0.21.10
+
+### Patch Changes
+
+- @hot-updater/core@0.21.10
+
+## 0.21.9
+
+### Patch Changes
+
+- @hot-updater/core@0.21.9
+
 ## 0.21.8
 
 ### Patch Changes