+++ plugins/plugin-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @hot-updater/plugin-core
 
+## 0.21.10
+
+### Patch Changes
+
+- @hot-updater/core@0.21.10
+
+## 0.21.9
+
+### Patch Changes
+
+- aa399a6: chore: deps picocolors
+  - @hot-updater/core@0.21.9
+
 ## 0.21.8
 
 ### Patch Changes